feat: add force-api-refresh option to the api push command in order to support refreshing an API (for example, on resolver code changes)
